# src/main/resources/swagger/odg-api.yaml
swagger: "2.0"
info:
  description: "Service for components at OMS to generate documents"
  version: "1.0.0"
  title: "OrchardDocumentGeneratorApi"
tags:
  - name: "ODG"
    externalDocs:
      description: "Find out more"
      url: "https://github.com/adamsiedlecki/ODG"
schemes:
  - "https"
  - "http"
paths:
  /jfree-chart/xy-plot:
    post:
      tags:
        - "ODG"
      summary: "Return XY chart"
      operationId: "createXyChart"
      consumes:
        - "application/json"
      produces:
        - "image/jpeg"
      parameters:
        - in: "body"
          name: "CreateChartInput"
          description: Body of the chart request
          required: true
          schema:
            $ref: "#/definitions/CreateChartInput"
      responses:
        "200":
          description: "OK"
          schema:
            type: file
        "400":
          description: "Bad request"
        "500":
          description: "Unexpected error"

  /jfree-chart/bar/chart:
    post:
      tags:
        - "ODG"
      summary: "Return bar chart"
      operationId: "createBarChart"
      consumes:
        - "application/json"
      produces:
        - "image/jpeg"
      parameters:
        - in: "body"
          name: "CreateBarChartInput"
          description: Body of the chart request
          required: true
          schema:
            $ref: "#/definitions/CreateBarChartInput"
      responses:
        "200":
          description: "OK"
          schema:
            type: file
        "400":
          description: "Bad request"
        "500":
          description: "Unexpected error"

definitions:
  CreateChartInput:
    type: object
    properties:
      chartTitle:
        type: string
      valuesLabel:
        type: string
      timeLabel:
        type: string
      widthPixels:
        type: integer
        default: 600
      heightPixels:
        type: integer
        default: 500
      areItemLabelsVisible:
        type: boolean
      redValueMarkerLineLevel:
        type: number
      maxMinutesConnectingLines:
        description: max amount of minutes between two values that still results in a line that connects it on a chart
        type: integer
        default: 120
      valueList:
        type: array
        items:
          $ref: "#/definitions/PresentableOnChart"
    required:
      - valueList

  PresentableOnChart:
    type: object
    properties:
      time:
        type: string
        format: date-time
      value:
        type: number
      groupName:
        type: string
    required:
      - time
      - value
      - groupName

  CreateBarChartInput:
    type: object
    properties:
      chartTitle:
        type: string
      valuesLabel:
        type: string
      categoriesLabel:
        type: string
      widthPixels:
        type: integer
        default: 600
      heightPixels:
        type: integer
        default: 500
      areItemLabelsVisible:
        type: boolean
      maxValueMarkerText:
        type: string
      valueList:
        type: array
        items:
          $ref: "#/definitions/PresentableOnBarChart"
    required:
      - valueList

  PresentableOnBarChart:
    type: object
    properties:
      value:
        type: number
      categoryName:
        type: string
      subCategoryName:
        type: string
    required:
      - value
      - categoryName
      - subCategoryName